daniele_marco/daniele/mnt/user-data/outputs/contractiq/parser_service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pdfplumber==0.10.3
python-docx==1.1.0
